+++ src/streamdiffusion/acceleration/tensorrt/export_wrappers/unet_ipadapter_export.py
@@ -2,7 +2,7 @@ import torch
 from diffusers import UNet2DConditionModel
 from typing import Optional, Dict, Any, List
 
-from ...model_detection import detect_model, detect_model_from_diffusers_unet
+from ....model_detection import detect_model, detect_model_from_diffusers_unet
 
 class IPAdapterUNetExportWrapper(torch.nn.Module):
     """
